# WorkCard-Lifecycle/openapi/openapi.json
{
  "components": {
    "schemas": {
      "BootstrapResponse": {
        "properties": {
          "csrfToken": {
            "title": "Csrftoken",
            "type": "string"
          },
          "expiresInSeconds": {
            "title": "Expiresinseconds",
            "type": "integer"
          }
        },
        "required": [
          "csrfToken",
          "expiresInSeconds"
        ],
        "title": "BootstrapResponse",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "IdentityListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IdentityResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "IdentityListResponse",
        "type": "object"
      },
      "IdentityResponse": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "PLANNER",
              "MASTER",
              "WORKER",
              "QUALITY_CONTROLLER",
              "ADMIN_AUDITOR"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "role"
        ],
        "title": "IdentityResponse",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Machine-readable error response shared by runtime and OpenAPI.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "traceId": {
            "title": "Traceid",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "traceId",
          "errors"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemFieldError": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "title": "ProblemFieldError",
        "type": "object"
      },
      "ReadinessResponse": {
        "properties": {
          "checks": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Checks",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks"
        ],
        "title": "ReadinessResponse",
        "type": "object"
      },
      "SelectIdentityRequest": {
        "additionalProperties": false,
        "properties": {
          "demoIdentityId": {
            "format": "uuid",
            "title": "Demoidentityid",
            "type": "string"
          }
        },
        "required": [
          "demoIdentityId"
        ],
        "title": "SelectIdentityRequest",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/IdentityResponse"
          },
          "csrfToken": {
            "title": "Csrftoken",
            "type": "string"
          },
          "expiresInSeconds": {
            "title": "Expiresinseconds",
            "type": "integer"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          }
        },
        "required": [
          "actor",
          "permissions",
          "csrfToken",
          "expiresInSeconds"
        ],
        "title": "SessionResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Production Work Card Workflow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/demo-identities": {
      "get": {
        "operationId": "list_demo_identities_api_v1_demo_identities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          }
        },
        "summary": "List Demo Identities",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session": {
      "delete": {
        "operationId": "delete_session_api_v1_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "session"
        ]
      },
      "get": {
        "operationId": "get_session_api_v1_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          }
        },
        "summary": "Get Session",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/bootstrap": {
      "get": {
        "operationId": "bootstrap_session_api_v1_session_bootstrap_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          }
        },
        "summary": "Bootstrap Session",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/demo": {
      "put": {
        "operationId": "select_demo_identity_api_v1_session_demo_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          }
        },
        "summary": "Select Demo Identity",
        "tags": [
          "session"
        ]
      }
    },
    "/health/live": {
      "get": {
        "operationId": "health_live_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          }
        },
        "summary": "Health Live",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "health_ready_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem Details error response"
          }
        },
        "summary": "Health Ready",
        "tags": [
          "health"
        ]
      }
    }
  }
}